backend/requirements.txt
# Backend — Auto Workshop System
# Python 3.11+

fastapi==0.134.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0

sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9

python-jose[cryptography]>=3.3.0
bcrypt>=4.0

python-multipart>=0.0.6
email-validator>=2.0

# Тесты
pytest>=7.0
httpx>=0.25
